+++ packages/bezier-react/src/components/Forms/Inputs/TextArea/TextArea.styled.ts
@@ -50,13 +50,7 @@ interface TextAreaAutoSizeBaseProps extends InterpolationProps {
   readOnly: boolean
 }
 
-/**
- * FIXME: https://github.com/rollup/plugins/issues/872
- * @rollup/plugin-commonjs 의 버그로 인해 default export 가 namespace 그 자체로 계산됨.
- * commonjs 상황을 위해 '.default' 를 추가함.
- */
-// @ts-ignore
-const TextAreaAutoSizeBase = styled(TextareaAutosize.default ?? TextareaAutosize)<TextAreaAutoSizeBaseProps>`
+const TextAreaAutoSizeBase = styled(TextareaAutosize)<TextAreaAutoSizeBaseProps>`
   box-sizing: border-box;
   width: 100%;
   padding: 8px 12px;