+++ packages/bezier-icons/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @channel.io/bezier-icons
 
+## 0.3.0
+
+### Minor Changes
+
+- Update icons ([#1426](https://github.com/channel-io/bezier-react/pull/1426)) by @Jamie-channel
+
+  - **NEW**: ImageQuality
+  - **EDIT**: CheckVerificationFilled, Download, Upload, VideocamOffFilled
+
+### Patch Changes
+
+- Improve the way module resolves. ([#1424](https://github.com/channel-io/bezier-react/pull/1424)) by @sungik-choi
+
+## 0.2.0
+
+### Minor Changes
+
+- Change the identifier of the bezier icon to a string literal instead of a symbol ([#1411](https://github.com/channel-io/bezier-react/pull/1411)) by @sungik-choi
+
 ## 0.1.1
 
 ### Patch Changes